move out autoAuth logic from ModalContainer
